Move diagnostic test to its own target, rename usage of diagnoseshards
