+++ cloud/filestore/tests/diagnostic_test/nfs-storage.txt
@@ -1,0 +1,5 @@
+LargeDeletionMarkersEnabled: true
+MaxFileBlocks: 536870912
+AutomaticShardCreationEnabled: true
+ShardAllocationUnit: 1073741824
+AutomaticallyCreatedShardSize: 1073741824
